Remove backend server for hostos-rollout cmd CLI
