+++ rs/cli/src/operations/mod.rs
@@ -1,0 +1,1 @@
+pub mod hostos_rollout;
