+++ rs/ic-management-backend/Cargo.toml
@@ -9,7 +9,6 @@ documentation.workspace = true
 [dependencies]
 actix-web = { workspace = true }
 anyhow = { workspace = true }
-async-recursion = { workspace = true }
 async-trait = { workspace = true }
 backon = { workspace = true }
 candid = { workspace = true }
